修改RID 8->4 字节 page 128K
10W
index 1  	4 -> 2.8
index 2 	4.2 -> 2.5
table_data  14.7

+++ src/observer/defs.h
@@ -14,5 +14,5 @@ See the Mulan PSL v2 for more details. */
 
 #pragma once
 
-using PageNum = int32_t;
-using SlotNum = int32_t;
+using PageNum = int16_t;
+using SlotNum = int16_t;
